requirements.txt
matplotlib==3.9.1
Pillow
Requests==2.32.3
streamlit==1.32.0
torch==2.3.1
torchvision
torchaudio
seaborn
